Moved ingest models into separate file
